Move some util classes to polaris-service-common as test fixtures
