+++ quarkus/service/build.gradle.kts
@@ -90,6 +90,7 @@ dependencies {
   testFixturesApi(project(":polaris-tests"))
 
   testImplementation(project(":polaris-api-management-model"))
+  testImplementation(testFixtures(project(":polaris-service-common")))
 
   testImplementation("org.apache.iceberg:iceberg-api:${libs.versions.iceberg.get()}:tests")
   testImplementation("org.apache.iceberg:iceberg-core:${libs.versions.iceberg.get()}:tests")
