+++ service/common/build.gradle.kts
@@ -19,6 +19,7 @@
 
 plugins {
   id("polaris-server")
+  id("java-test-fixtures")
   alias(libs.plugins.jandex)
 }
 
@@ -91,6 +92,27 @@ dependencies {
   testImplementation(libs.assertj.core)
   testImplementation(libs.mockito.core)
   testRuntimeOnly("org.junit.platform:junit-platform-launcher")
+
+  testFixturesImplementation(project(":polaris-core"))
+  testFixturesImplementation(project(":polaris-api-management-service"))
+  testFixturesImplementation(project(":polaris-api-iceberg-service"))
+
+  testFixturesImplementation(libs.jakarta.enterprise.cdi.api)
+  testFixturesImplementation(libs.jakarta.annotation.api)
+  testFixturesImplementation(libs.jakarta.inject.api)
+  testFixturesImplementation(libs.jakarta.ws.rs.api)
+
+  testFixturesImplementation(platform(libs.iceberg.bom))
+  testFixturesImplementation("org.apache.iceberg:iceberg-api")
+  testFixturesImplementation("org.apache.iceberg:iceberg-core")
+  testFixturesImplementation("org.apache.iceberg:iceberg-aws")
+
+  testFixturesImplementation(platform(libs.google.cloud.storage.bom))
+  testFixturesImplementation("com.google.cloud:google-cloud-storage")
+  testFixturesImplementation(platform(libs.awssdk.bom))
+  testFixturesImplementation("software.amazon.awssdk:sts")
+  testFixturesImplementation("software.amazon.awssdk:iam-policy-builder")
+  testFixturesImplementation("software.amazon.awssdk:s3")
 }
 
 tasks.named("javadoc") { dependsOn("jandex") }
